feat: add import/export model to jobs
